+++ juniper_codegen/src/graphql_union/mod.rs
@@ -3,13 +3,16 @@ pub mod derive;
 
 use std::collections::HashMap;
 
+use proc_macro2::{Span, TokenStream};
+use quote::quote;
 use syn::{
     parse::{Parse, ParseStream},
+    parse_quote,
     spanned::Spanned as _,
 };
 
 use crate::util::{
-    filter_graphql_attrs, get_doc_comment, span_container::SpanContainer, OptionExt as _,
+    filter_graphql_attrs, get_doc_comment, span_container::SpanContainer, Mode, OptionExt as _,
 };
 
 /// Available metadata behind `#[graphql]` (or `#[graphql_union]`) attribute when generating code
@@ -280,5 +283,246 @@ impl UnionVariantMeta {
         filter_graphql_attrs(attrs)
             .map(|attr| attr.parse_args())
             .try_fold(Self::default(), |prev, curr| prev.try_merge(curr?))
+    }
+}
+
+struct UnionVariantDefinition {
+    pub ty: syn::Type,
+    pub resolver_code: syn::Expr,
+    pub resolver_check: syn::Expr,
+    pub enum_path: Option<TokenStream>,
+    pub span: Span,
+}
+
+struct UnionDefinition {
+    pub name: String,
+    pub ty: syn::Type,
+    pub description: Option<String>,
+    pub context: Option<syn::Type>,
+    pub scalar: Option<syn::Type>,
+    pub generics: syn::Generics,
+    pub variants: Vec<UnionVariantDefinition>,
+    pub span: Span,
+    pub mode: Mode,
+}
+
+impl UnionDefinition {
+    pub fn into_tokens(self) -> TokenStream {
+        let crate_path = self.mode.crate_path();
+
+        let name = &self.name;
+        let ty = &self.ty;
+
+        let context = self
+            .context
+            .as_ref()
+            .map(|ctx| quote! { #ctx })
+            .unwrap_or_else(|| quote! { () });
+
+        let scalar = self
+            .scalar
+            .as_ref()
+            .map(|scl| quote! { #scl })
+            .unwrap_or_else(|| quote! { __S });
+        let default_scalar = self
+            .scalar
+            .as_ref()
+            .map(|scl| quote! { #scl })
+            .unwrap_or_else(|| quote! { #crate_path::DefaultScalarValue });
+
+        let description = self
+            .description
+            .as_ref()
+            .map(|desc| quote! { .description(#desc) });
+
+        let var_types: Vec<_> = self.variants.iter().map(|var| &var.ty).collect();
+
+        let match_names = self.variants.iter().map(|var| {
+            let var_ty = &var.ty;
+            let var_check = &var.resolver_check;
+            quote! {
+                if #var_check {
+                    return <#var_ty as #crate_path::GraphQLType<#scalar>>::name(&())
+                        .unwrap().to_string();
+                }
+            }
+        });
+
+        let match_resolves: Vec<_> = self.variants.iter().map(|var| &var.resolver_code).collect();
+        let resolve_into_type = self.variants.iter().zip(match_resolves.iter()).map(|(var, expr)| {
+            let var_ty = &var.ty;
+
+            let get_name = quote! { (<#var_ty as #crate_path::GraphQLType<#scalar>>::name(&())) };
+            quote! {
+                if type_name == #get_name.unwrap() {
+                    return #crate_path::IntoResolvable::into(
+                        { #expr },
+                        executor.context()
+                    )
+                    .and_then(|res| match res {
+                        Some((ctx, r)) => executor.replaced_context(ctx).resolve_with_ctx(&(), &r),
+                        None => Ok(#crate_path::Value::null()),
+                    });
+                }
+            }
+        });
+        let resolve_into_type_async =
+            self.variants
+                .iter()
+                .zip(match_resolves.iter())
+                .map(|(var, expr)| {
+                    let var_ty = &var.ty;
+
+                    let get_name =
+                        quote! { (<#var_ty as #crate_path::GraphQLType<#scalar>>::name(&())) };
+                    quote! {
+                        if type_name == #get_name.unwrap() {
+                            let res = #crate_path::IntoResolvable::into(
+                                { #expr },
+                                executor.context()
+                            );
+                            return #crate_path::futures::future::FutureExt::boxed(async move {
+                                match res? {
+                                    Some((ctx, r)) => {
+                                        let subexec = executor.replaced_context(ctx);
+                                        subexec.resolve_with_ctx_async(&(), &r).await
+                                    },
+                                    None => Ok(#crate_path::Value::null()),
+                                }
+                            });
+                        }
+                    }
+                });
+
+        let (impl_generics, ty_generics, _) = self.generics.split_for_impl();
+        let mut ext_generics = self.generics.clone();
+        if self.scalar.is_none() {
+            ext_generics.params.push(parse_quote! { #scalar });
+            ext_generics
+                .where_clause
+                .get_or_insert_with(|| parse_quote! { where })
+                .predicates
+                .push(parse_quote! { #scalar: #crate_path::ScalarValue });
+        }
+        let (ext_impl_generics, _, where_clause) = ext_generics.split_for_impl();
+
+        let mut where_async = where_clause
+            .cloned()
+            .unwrap_or_else(|| parse_quote! { where });
+        where_async
+            .predicates
+            .push(parse_quote! { Self: Send + Sync });
+        if self.scalar.is_none() {
+            where_async
+                .predicates
+                .push(parse_quote! { #scalar: Send + Sync });
+        }
+
+        let type_impl = quote! {
+            #[automatically_derived]
+            impl#ext_impl_generics #crate_path::GraphQLType<#scalar> for #ty#ty_generics
+                #where_clause
+            {
+                type Context = #context;
+                type TypeInfo = ();
+
+                fn name(_ : &Self::TypeInfo) -> Option<&str> {
+                    Some(#name)
+                }
+
+                fn meta<'r>(
+                    info: &Self::TypeInfo,
+                    registry: &mut #crate_path::Registry<'r, #scalar>
+                ) -> #crate_path::meta::MetaType<'r, #scalar>
+                where #scalar: 'r,
+                {
+                    let types = &[
+                        #( registry.get_type::<&#var_types>(&(())), )*
+                    ];
+                    registry.build_union_type::<#ty#ty_generics>(info, types)
+                    #description
+                    .into_meta()
+                }
+
+                fn concrete_type_name(
+                    &self,
+                    context: &Self::Context,
+                    _: &Self::TypeInfo,
+                ) -> String {
+                    #( #match_names )*
+                    panic!(
+                        "GraphQL union {} cannot be resolved into any of its variants in its \
+                         current state",
+                        #name,
+                    );
+                }
+
+                fn resolve_into_type(
+                    &self,
+                    _: &Self::TypeInfo,
+                    type_name: &str,
+                    _: Option<&[#crate_path::Selection<#scalar>]>,
+                    executor: &#crate_path::Executor<Self::Context, #scalar>,
+                ) -> #crate_path::ExecutionResult<#scalar> {
+                    let context = executor.context();
+                    #( #resolve_into_type )*
+                    panic!(
+                        "Concrete type {} is not handled by instance resolvers on GraphQL union {}",
+                        type_name, #name,
+                    );
+                }
+            }
+        };
+
+        let async_type_impl = quote! {
+            #[automatically_derived]
+            impl#ext_impl_generics #crate_path::GraphQLTypeAsync<#scalar> for #ty#ty_generics
+                #where_async
+            {
+                fn resolve_into_type_async<'b>(
+                    &'b self,
+                    _: &'b Self::TypeInfo,
+                    type_name: &str,
+                    _: Option<&'b [#crate_path::Selection<'b, #scalar>]>,
+                    executor: &'b #crate_path::Executor<'b, 'b, Self::Context, #scalar>
+                ) -> #crate_path::BoxFuture<'b, #crate_path::ExecutionResult<#scalar>> {
+                    let context = executor.context();
+                    #( #resolve_into_type_async )*
+                    panic!(
+                        "Concrete type {} is not handled by instance resolvers on GraphQL union {}",
+                        type_name, #name,
+                    );
+                }
+            }
+        };
+
+        let output_type_impl = quote! {
+            #[automatically_derived]
+            impl#ext_impl_generics #crate_path::marker::IsOutputType<#scalar> for #ty#ty_generics
+                #where_clause
+            {
+                fn mark() {
+                    #( <#var_types as #crate_path::marker::GraphQLObjectType<#scalar>>::mark(); )*
+                }
+            }
+        };
+
+        let union_impl = quote! {
+            #[automatically_derived]
+            impl#impl_generics #crate_path::marker::GraphQLUnion for #ty#ty_generics {
+                fn mark() {
+                    #( <#var_types as #crate_path::marker::GraphQLObjectType<
+                        #default_scalar,
+                    >>::mark(); )*
+                }
+            }
+        };
+
+        quote! {
+            #union_impl
+            #output_type_impl
+            #type_impl
+            #async_type_impl
+        }
     }
 }
